Add pointer type (#5)
* ensure all variables are stack allocated with "StackSlot"s

* allow compilation of stack based variables

* added *.exe* to .gitignore

* allow println and print builtins to take a uint64_t to print pointers

* ensure that stack slot allocation uses correct variable sizes

* added ptr_type.wellick example

* added pointer operators (&, *) to parser

* added compilation code to facilitate creating/ dereferencing pointers

+++ builtins.c
@@ -1,5 +1,6 @@
 #include <stdint.h>
 #include <stdio.h>
+#include <inttypes.h>
 
 // A collection of C functions that are linked to Wellick
 // and can be called from Wellick code. These functions act
@@ -9,13 +10,13 @@
 // So all integer functions use uint32_t in their parameters and return types
 
 // Print a string to stdout
-int print(uint32_t inp) {
-    return printf("%d", inp);
+int print(uint64_t inp) {
+    return printf("%" PRId64, inp);
 }
 
 // Print a string to stdout with a newline
-int println(uint32_t inp) {
-    return printf("%d\n", inp);
+int println(uint64_t inp) {
+    return printf("%" PRId64 "\n", inp);
 }
 
 // Add two integers, return the result
